created a set of passing tests and fixed the stackoverflow in some tests by adding a maxStackDepth configuration parameter, setting it to 2 in case we are running in a dalvik vm.

+++ db4ounit.extensions/src/db4ounit/extensions/fixtures/AbstractFileBasedDb4oFixture.java
@@ -33,7 +33,9 @@ public abstract class AbstractFileBasedDb4oFixture extends AbstractSoloDb4oFixtu
 	}
 
 	protected void doClean() {
-		_databaseFile.delete();
+		if (_databaseFile.exists()) {
+			_databaseFile.delete();
+		}
 	}
 
 }
